start new branch matrixOutput to get A-matrix output of he3sim add modified he3sim.cpp and matep.hpp modifcations of he3sim.cpp: real_t double; A.write() added into scaling_sim::write_positions() modification of matep.hpp real_t double

+++ matep.hpp
@@ -19,7 +19,7 @@
 #include <vector>
 
 
-using real_t = float;
+using real_t = double; // or float
 
 class Matep {
 public:
